+++ Presentation.Web/Controllers/External/V2/OrganizationV2Controller.cs
@@ -7,6 +7,8 @@ using Core.DomainModel.Organization;
 using Presentation.Web.Extensions;
 using Presentation.Web.Infrastructure.Attributes;
 using Presentation.Web.Models.External.V2;
+using Presentation.Web.Models.External.V2.Request;
+using Presentation.Web.Models.External.V2.Response;
 using Swashbuckle.Swagger.Annotations;
 
 namespace Presentation.Web.Controllers.External.V2
