+++ Presentation.Web/Extensions/QueryablePaginationExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using Core.ApplicationServices.Shared;
 using Presentation.Web.Models.External.V2;
+using Presentation.Web.Models.External.V2.Request;
 
 namespace Presentation.Web.Extensions
 {
