+++ Presentation.Web/Models/External/V2/Request/DeactivationReasonRequestDTO.cs
@@ -1,13 +1,13 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Presentation.Web.Models.External.V2
+namespace Presentation.Web.Models.External.V2.Request
 {
     public class DeactivationReasonRequestDTO
     {
         /// <summary>
         /// Reason for deactivation
         /// </summary>
-        [Required]
+        [Required(AllowEmptyStrings = false)]
         public string DeactivationReason { get; set; }
     }
 }
