+++ Presentation.Web/Models/External/V2/Request/StandardPaginationQuery.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using Core.ApplicationServices.Shared;
 
-namespace Presentation.Web.Models.External.V2
+namespace Presentation.Web.Models.External.V2.Request
 {
     public class StandardPaginationQuery
     {
