+++ Presentation.Web/Models/External/V2/Response/IdentityNamePairResponseDTO.cs
@@ -1,18 +1,18 @@
 ﻿using System;
 
-namespace Presentation.Web.Models.External.V2
+namespace Presentation.Web.Models.External.V2.Response
 {
     public class IdentityNamePairResponseDTO
     {
         /// <summary>
-        /// UUID
+        /// UUID which is unique within collection of entities of the same type
         /// </summary>
-        public Guid Uuid { get; set; }
+        public Guid Uuid { get; }
 
         /// <summary>
         /// Name
         /// </summary>
-        public string Name { get; set; }
+        public string Name { get; }
 
         public IdentityNamePairResponseDTO(Guid uuid, string name)
         {
