+++ Presentation.Web/Models/External/V2/Response/ItSystemResponseDTO.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Presentation.Web.Models.External.V2
+namespace Presentation.Web.Models.External.V2.Response
 {
     public class ItSystemResponseDTO
     {
