+++ Presentation.Web/Models/External/V2/Response/OrganizationResponseDTO.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Presentation.Web.Models.External.V2
+namespace Presentation.Web.Models.External.V2.Response
 {
     public class OrganizationResponseDTO: IdentityNamePairResponseDTO
     {
