+++ Presentation.Web/Models/External/V2/Response/RecommendedArchiveDutyResponseDTO.cs
@@ -1,4 +1,4 @@
-﻿namespace Presentation.Web.Models.External.V2
+﻿namespace Presentation.Web.Models.External.V2.Response
 {
     public class RecommendedArchiveDutyResponseDTO
     {
