+++ Presentation.Web/Presentation.Web.csproj
@@ -348,19 +348,19 @@
     <Compile Include="Extensions\QueryablePaginationExtensions.cs" />
     <Compile Include="Infrastructure\Filters\AdvisSendFailureFilter.cs" />
     <Compile Include="Infrastructure\KitosConstants.cs" />
-    <Compile Include="Models\External\V2\StandardPaginationQuery.cs" />
+    <Compile Include="Models\External\V2\Request\StandardPaginationQuery.cs" />
     <Compile Include="Models\UserNotification\UserNotificationDTO.cs" />
     <Compile Include="Models\KendoOrganizationalConfigurationDTO.cs" />
     <Compile Include="Controllers\External\V2\ExternalBaseController.cs" />
     <Compile Include="Controllers\External\V2\OrganizationV2Controller.cs" />
     <Compile Include="Helpers\StringHelpers.cs" />
-    <Compile Include="Models\External\V2\ItSystemStakeholderResponseDTO.cs" />
+    <Compile Include="Models\External\V2\Response\ItSystemStakeholderResponseDTO.cs" />
     <Compile Include="Controllers\External\V2\GeneralStakeholderController.cs" />
-    <Compile Include="Models\External\V2\ItInterfaceResponseDTO.cs" />
-    <Compile Include="Models\External\V2\ItSystemResponseDTO.cs" />
-    <Compile Include="Models\External\V2\ItInterfaceRequestDTO.cs" />
+    <Compile Include="Models\External\V2\Response\ItInterfaceResponseDTO.cs" />
+    <Compile Include="Models\External\V2\Response\ItSystemResponseDTO.cs" />
+    <Compile Include="Models\External\V2\Request\ItInterfaceRequestDTO.cs" />
     <Compile Include="Controllers\External\V2\RightsHolderItInterfaceController.cs" />
-    <Compile Include="Models\External\V2\DeactivationReasonRequestDTO.cs" />
+    <Compile Include="Models\External\V2\Request\DeactivationReasonRequestDTO.cs" />
     <Compile Include="Controllers\External\V2\GeneralController.cs" />
     <Compile Include="Controllers\External\V2\RightsHolderItSystemController.cs" />
     <Compile Include="Controllers\API\ItSystemUsageOptionsController.cs" />
@@ -429,10 +429,10 @@
     <Compile Include="Models\Csv\CsvColumnDefinition.cs" />
     <Compile Include="Models\Csv\CsvColumnIdentity.cs" />
     <Compile Include="Models\Csv\CsvResponseBuilder.cs" />
-    <Compile Include="Models\External\V2\IdentityNamePairResponseDTO.cs" />
-    <Compile Include="Models\External\V2\ItSystemRequestDTO.cs" />
-    <Compile Include="Models\External\V2\OrganizationResponseDTO.cs" />
-    <Compile Include="Models\External\V2\RecommendedArchiveDutyResponseDTO.cs" />
+    <Compile Include="Models\External\V2\Response\IdentityNamePairResponseDTO.cs" />
+    <Compile Include="Models\External\V2\Request\ItSystemRequestDTO.cs" />
+    <Compile Include="Models\External\V2\Response\OrganizationResponseDTO.cs" />
+    <Compile Include="Models\External\V2\Response\RecommendedArchiveDutyResponseDTO.cs" />
     <Compile Include="Models\FeatureToggle\TemporaryFeature.cs" />
     <Compile Include="Models\GDPR\AssignRoleDTO.cs" />
     <Compile Include="Models\GDPR\CreateDataProcessingRegistrationDTO.cs" />
